requirements.txt
ultralytics==8.0.202
torch==1.13.0a0+git7c98e70
numpy==1.24.4
opencv-python==4.8.0.76
supervision==0.11.1
pyserial==3.4
